refactor(utils): move CarNamesBuilder and Validator to utils
기존에 model 패키지에서 작성하였던 CarNamesBuilder, Validator 클래스를 utils 패키지로 따로 분리하였습니다.
